+++ src/ModelContextProtocol/Hosting/McpServerSingleSessionHostedService.cs
@@ -1,0 +1,13 @@
+﻿using Microsoft.Extensions.Hosting;
+using ModelContextProtocol.Server;
+
+namespace ModelContextProtocol.Hosting;
+
+/// <summary>
+/// Hosted service for a single-session (i.e stdio) MCP server.
+/// </summary>
+internal class McpServerSingleSessionHostedService(IMcpServer session) : BackgroundService
+{
+    /// <inheritdoc />
+    protected override Task ExecuteAsync(CancellationToken stoppingToken) => session.RunAsync(stoppingToken);
+}
