+++ src/ModelContextProtocol/Server/McpServer.cs
@@ -12,31 +12,67 @@ namespace ModelContextProtocol.Server;
 /// <inheritdoc />
 internal sealed class McpServer : McpJsonRpcEndpoint, IMcpServer
 {
-    private readonly string _serverDescription;
+    private readonly IServerTransport? _serverTransport;
     private readonly EventHandler? _toolsChangedDelegate;
 
-    private volatile bool _ran;
+    private ITransport? _sessionTransport;
+    private string _endpointName;
 
     /// <summary>
     /// Creates a new instance of <see cref="McpServer"/>.
     /// </summary>
-    /// <param name="transport">Transport to use for the server</param>
-    /// <param name="options">Configuration options for this server, including capabilities. 
-    /// Make sure to accurately reflect exactly what capabilities the server supports and does not support.</param>    
+    /// <param name="serverTransport">Transport to use for the server that is ready to accept new sessions asynchronously.</param>
+    /// <param name="options">Configuration options for this server, including capabilities.
+    /// Make sure to accurately reflect exactly what capabilities the server supports and does not support.</param>
+    /// <param name="loggerFactory">Logger factory to use for logging</param>
+    /// <param name="serviceProvider">Optional service provider to use for dependency injection</param>
+    /// <exception cref="McpServerException"></exception>
+    public McpServer(IServerTransport serverTransport, McpServerOptions options, ILoggerFactory? loggerFactory, IServiceProvider? serviceProvider)
+        : this(options, loggerFactory, serviceProvider)
+    {
+        Throw.IfNull(serverTransport);
+
+        _serverTransport = serverTransport;
+    }
+
+    /// <summary>
+    /// Creates a new instance of <see cref="McpServer"/>.
+    /// </summary>
+    /// <param name="transport">Transport to use for the server representing an already-established session.</param>
+    /// <param name="options">Configuration options for this server, including capabilities.
+    /// Make sure to accurately reflect exactly what capabilities the server supports and does not support.</param>
     /// <param name="loggerFactory">Logger factory to use for logging</param>
     /// <param name="serviceProvider">Optional service provider to use for dependency injection</param>
     /// <exception cref="McpServerException"></exception>
     public McpServer(ITransport transport, McpServerOptions options, ILoggerFactory? loggerFactory, IServiceProvider? serviceProvider)
-        : base(transport, loggerFactory)
+        : this(options, loggerFactory, serviceProvider)
+    {
+        Throw.IfNull(transport);
+
+        _sessionTransport = transport;
+        InitializeSession(transport);
+    }
+
+    /// <summary>
+    /// Creates a new instance of <see cref="McpServer"/>.
+    /// </summary>
+    /// <param name="options">Configuration options for this server, including capabilities. 
+    /// Make sure to accurately reflect exactly what capabilities the server supports and does not support.</param>
+    /// <param name="loggerFactory">Logger factory to use for logging</param>
+    /// <param name="serviceProvider">Optional service provider to use for dependency injection</param>
+    /// <exception cref="McpServerException"></exception>
+    private McpServer(McpServerOptions options, ILoggerFactory? loggerFactory, IServiceProvider? serviceProvider)
+        : base(loggerFactory)
     {
         Throw.IfNull(options);
 
         ServerOptions = options;
         Services = serviceProvider;
-        _serverDescription = $"{options.ServerInfo.Name} {options.ServerInfo.Version}";
+        _endpointName = $"Server ({options.ServerInfo.Name} {options.ServerInfo.Version})";
+
         _toolsChangedDelegate = delegate
         {
-            _ = SendMessageAsync(new JsonRpcNotification()
+            _ = _session?.SendMessageAsync(new JsonRpcNotification()
             {
                 Method = NotificationMethods.ToolListChangedNotification,
             });
@@ -77,60 +113,63 @@ internal sealed class McpServer : McpJsonRpcEndpoint, IMcpServer
     public IServiceProvider? Services { get; }
 
     /// <inheritdoc />
-    public override string EndpointName =>
-        $"Server ({_serverDescription}), Client ({ClientInfo?.Name} {ClientInfo?.Version})";
+    public override string EndpointName => _endpointName;
 
-    /// <inheritdoc />
-    public async Task RunAsync(Func<CancellationToken, Task>? onInitialized = null, CancellationToken cancellationToken = default)
+    public async Task AcceptSessionAsync(CancellationToken cancellationToken = default)
     {
-        if (_ran)
-        {
-            _logger.ServerAlreadyInitializing(EndpointName);
-            throw new InvalidOperationException("Server is already running.");
-        }
-        _ran = true;
+        // Below is effectively an assertion. The McpServerFactory should only use this with the IServerTransport constructor.
+        Throw.IfNull(_serverTransport);
 
         try
         {
-            CancellationTokenSource = CancellationTokenSource.CreateLinkedTokenSource(cancellationToken);
+            _sessionTransport = await _serverTransport.AcceptAsync(cancellationToken).ConfigureAwait(false);
 
-            IServerTransport? serverTransport = Transport as IServerTransport;
-            if (serverTransport is not null)
+            if (_sessionTransport is null)
             {
-                // Start listening for messages
-                await serverTransport.StartListeningAsync(CancellationTokenSource.Token).ConfigureAwait(false);
+                throw new McpServerException("The server transport closed before a client started a new session.");
             }
 
-            // Start processing messages
-            MessageProcessingTask = ProcessMessagesAsync(CancellationTokenSource.Token);
-
-            // Invoke any post-initialization logic.
-            if (onInitialized is not null)
-            {
-                await onInitialized(CancellationTokenSource.Token).ConfigureAwait(false);
-            }
-
-            // Wait for transport completion.
-            if (serverTransport is not null)
-            {
-                await serverTransport.Completion;
-            }
+            InitializeSession(_sessionTransport);
         }
         catch (Exception e)
         {
             _logger.ServerInitializationError(EndpointName, e);
-            await CleanupAsync().ConfigureAwait(false);
             throw;
         }
     }
 
-    protected override Task CleanupAsync()
+    /// <inheritdoc />
+    public async Task RunAsync(CancellationToken cancellationToken = default)
+    {
+        // Below is effectively an assertion. The McpServerFactory should not return before the _transport is initialized.
+        Throw.IfNull(_session);
+
+        try
+        {
+            // Start processing messages
+            StartSession(fullSessionCancellationToken: cancellationToken);
+            await MessageProcessingTask.ConfigureAwait(false);
+        }
+        finally
+        {
+            await DisposeAsync().ConfigureAwait(false);
+        }
+    }
+
+    public override async ValueTask DisposeAsync()
     {
         if (ServerOptions.Capabilities?.Tools?.ToolCollection is { } tools)
         {
             tools.Changed -= _toolsChangedDelegate;
         }
-        return base.CleanupAsync();
+
+        await base.DisposeAsync().ConfigureAwait(false);
+
+        if (_serverTransport is not null && _sessionTransport is not null)
+        {
+            // We created the _sessionTransport from the _serverTransport, so we own it.
+            await _sessionTransport.DisposeAsync().ConfigureAwait(false);
+        }
     }
 
     private void SetPingHandler()
@@ -146,6 +185,11 @@ internal sealed class McpServer : McpJsonRpcEndpoint, IMcpServer
             {
                 ClientCapabilities = request?.Capabilities ?? new();
                 ClientInfo = request?.ClientInfo;
+
+                // Use the ClientInfo to update the session EndpointName for logging.
+                _endpointName = $"{_endpointName}, Client ({ClientInfo?.Name} {ClientInfo?.Version})";
+                GetSessionOrThrow().EndpointName = EndpointName;
+
                 return Task.FromResult(new InitializeResult()
                 {
                     ProtocolVersion = options.ProtocolVersion,
