+++ tests/ModelContextProtocol.Tests/Server/McpServerTests.cs
@@ -1,6 +1,5 @@
 ﻿using Microsoft.Extensions.AI;
 using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Logging;
 using ModelContextProtocol.Client;
 using ModelContextProtocol.Protocol.Messages;
 using ModelContextProtocol.Protocol.Transport;
@@ -14,16 +13,14 @@ namespace ModelContextProtocol.Tests.Server;
 
 public class McpServerTests : LoggedTest
 {
-    private readonly Mock<IServerTransport> _serverTransport;
-    private readonly Mock<ILogger> _logger;
+    private readonly Mock<ITransport> _serverTransport;
     private readonly McpServerOptions _options;
     private readonly IServiceProvider _serviceProvider;
 
     public McpServerTests(ITestOutputHelper testOutputHelper)
         : base(testOutputHelper)
     {
-        _serverTransport = new Mock<IServerTransport>();
-        _logger = new Mock<ILogger>();
+        _serverTransport = new Mock<ITransport>();
         _options = CreateOptions();
         _serviceProvider = new ServiceCollection().BuildServiceProvider();
     }
@@ -88,42 +85,19 @@ public class McpServerTests : LoggedTest
     {
         // Arrange
         await using var server = McpServerFactory.Create(_serverTransport.Object, _options, LoggerFactory, _serviceProvider);
-        var task = server.RunAsync(cancellationToken: TestContext.Current.CancellationToken);
+        var runTask = server.RunAsync(TestContext.Current.CancellationToken);
 
         // Act & Assert
-        await Assert.ThrowsAsync<InvalidOperationException>(() => server.RunAsync(cancellationToken: TestContext.Current.CancellationToken));
+        await Assert.ThrowsAsync<InvalidOperationException>(() => server.RunAsync(TestContext.Current.CancellationToken));
 
-        await task;
-    }
-
-    [Fact]
-    public async Task RunAsync_ShouldStartListening()
-    {
-        // Arrange
-        await using var server = McpServerFactory.Create(_serverTransport.Object, _options, LoggerFactory, _serviceProvider);
-
-        // Act
-        await server.RunAsync(cancellationToken: TestContext.Current.CancellationToken);
-
-        // Assert
-        _serverTransport.Verify(t => t.StartListeningAsync(It.IsAny<CancellationToken>()), Times.Once);
-    }
-
-    [Fact]
-    public async Task RunAsync_Sets_Initialized_After_Transport_Responses_Initialized_Notification()
-    {
-        await using var transport = new TestServerTransport();
-        await using var server = McpServerFactory.Create(transport, _options, LoggerFactory, _serviceProvider);
-
-        await server.RunAsync(cancellationToken: TestContext.Current.CancellationToken);
-
-        // Send initialized notification
-        await transport.SendMessageAsync(new JsonRpcNotification
-            {
-                Method = "notifications/initialized"
-            }, TestContext.Current.CancellationToken);
-
-        await Task.Delay(50, TestContext.Current.CancellationToken);
+        try
+        {
+            await runTask;
+        }
+        catch (NullReferenceException)
+        {
+            // _serverTransport.Object returns a null MessageReader
+        }
     }
 
     [Fact]
@@ -147,7 +121,7 @@ public class McpServerTests : LoggedTest
         await using var server = McpServerFactory.Create(transport, _options, LoggerFactory, _serviceProvider);
         SetClientCapabilities(server, new ClientCapabilities { Sampling = new SamplingCapability() });
 
-        await server.RunAsync(cancellationToken: TestContext.Current.CancellationToken);
+        var runTask = server.RunAsync(TestContext.Current.CancellationToken);
 
         // Act
         var result = await server.RequestSamplingAsync(new CreateMessageRequestParams { Messages = [] }, CancellationToken.None);
@@ -156,6 +130,9 @@ public class McpServerTests : LoggedTest
         Assert.NotEmpty(transport.SentMessages);
         Assert.IsType<JsonRpcRequest>(transport.SentMessages[0]);
         Assert.Equal("sampling/createMessage", ((JsonRpcRequest)transport.SentMessages[0]).Method);
+
+        await transport.DisposeAsync();
+        await runTask;
     }
 
     [Fact]
@@ -176,7 +153,7 @@ public class McpServerTests : LoggedTest
         await using var transport = new TestServerTransport();
         await using var server = McpServerFactory.Create(transport, _options, LoggerFactory, _serviceProvider);
         SetClientCapabilities(server, new ClientCapabilities { Roots = new RootsCapability() });
-        await server.RunAsync(cancellationToken: TestContext.Current.CancellationToken);
+        var runTask = server.RunAsync(TestContext.Current.CancellationToken);
 
         // Act
         var result = await server.RequestRootsAsync(new ListRootsRequestParams(), CancellationToken.None);
@@ -186,6 +163,9 @@ public class McpServerTests : LoggedTest
         Assert.NotEmpty(transport.SentMessages);
         Assert.IsType<JsonRpcRequest>(transport.SentMessages[0]);
         Assert.Equal("roots/list", ((JsonRpcRequest)transport.SentMessages[0]).Method);
+
+        await transport.DisposeAsync();
+        await runTask;
     }
 
     [Fact]
@@ -540,7 +520,7 @@ public class McpServerTests : LoggedTest
 
         await using var server = McpServerFactory.Create(transport, options, LoggerFactory, _serviceProvider);
 
-        await server.RunAsync();
+        var runTask = server.RunAsync(TestContext.Current.CancellationToken);
 
         var receivedMessage = new TaskCompletionSource<JsonRpcResponse>();
 
@@ -563,6 +543,9 @@ public class McpServerTests : LoggedTest
         Assert.NotNull(response.Result);
 
         assertResult(response.Result);
+
+        await transport.DisposeAsync();
+        await runTask;
     }
 
     private async Task Throws_Exception_If_No_Handler_Assigned(ServerCapabilities serverCapabilities, string method, string expectedError)
@@ -660,7 +643,7 @@ public class McpServerTests : LoggedTest
             throw new NotImplementedException();
         public Task SendMessageAsync(IJsonRpcMessage message, CancellationToken cancellationToken = default) =>
             throw new NotImplementedException();
-        public Task RunAsync(Func<CancellationToken, Task>? onInitialized = null, CancellationToken cancellationToken = default) =>
+        public Task RunAsync(CancellationToken cancellationToken = default) =>
             throw new NotImplementedException();
     }
 }
